!XT (Audio) Added functions to set parameters and states globally on all objects. !XT (Audio) Executing and stopping triggers on the global object is now a system request.
Copied from Perforce
 Change: 1879250

+++ Code/CryEngine/CryAudioSystem/implementations/CryAudioImplSDLMixer/GlobalData.h
@@ -63,9 +63,9 @@ struct STriggerInfo final : public ITriggerInfo
 
 struct SPoolSizes final
 {
-	uint32 triggers = 0;
-	uint32 parameters = 0;
-	uint32 switchStates = 0;
+	uint16 triggers = 0;
+	uint16 parameters = 0;
+	uint16 switchStates = 0;
 };
 } // namespace SDL_mixer
 } // namespace Impl
